added basic command options
