+++ BedrockCore/src/main/java/de/cubenation/bedrock/core/annotation/ArgumentConstraints.java
@@ -1,0 +1,13 @@
+package de.cubenation.bedrock.core.annotation;
+
+import de.cubenation.bedrock.core.annotation.condition.DefaultCondition;
+
+import java.lang.annotation.*;
+
+@Documented
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.PARAMETER)
+public @interface ArgumentConstraints {
+
+    Class Condition() default DefaultCondition.class;
+}
