+++ BedrockCore/src/main/java/de/cubenation/bedrock/core/message/Messages.java
@@ -24,9 +24,9 @@ package de.cubenation.bedrock.core.message;
 
 import de.cubenation.bedrock.core.FoundationPlugin;
 import de.cubenation.bedrock.core.authorization.Permission;
-import de.cubenation.bedrock.core.command.AbstractCommand;
+import de.cubenation.bedrock.core.command.Command;
 import de.cubenation.bedrock.core.command.argument.Argument;
-import de.cubenation.bedrock.core.command.argument.KeyValueArgument;
+import de.cubenation.bedrock.core.command.argument.Option;
 import de.cubenation.bedrock.core.exception.LocalizationNotFoundException;
 import de.cubenation.bedrock.core.helper.LengthComparator;
 import de.cubenation.bedrock.core.service.colorscheme.ColorScheme;
@@ -248,7 +248,7 @@ public abstract class Messages {
      * @param command a class that is abstracted from AbstractCommand
      * @return the TextComponent with the help.
      */
-    public JsonMessage getHelpForSubCommand(BedrockChatSender sender, AbstractCommand command) {
+    public JsonMessage getHelpForSubCommand(BedrockChatSender sender, Command command) {
 
         // check for permission
         if (!command.hasPermission(sender)) {
@@ -356,17 +356,17 @@ public abstract class Messages {
              * In case the argument is an instanceof the KeyValueArgument class (which is kind of a
              * key-value command) we need to prepend the key
              */
-            if (argument instanceof KeyValueArgument) {
-                KeyValueArgument keyValueArgument = (KeyValueArgument) argument;
+            if (argument instanceof Option) {
+                Option option = (Option) argument;
 
-                String keyString = "[" + keyValueArgument.getKey() + (keyValueArgument.getKeyOnly() ? "]" : "");
+                String keyString = "[" + option.getKey() + (!option.hasParameter() ? "]" : "");
 
                 helpJson.addExtra(BedrockJson.JsonWithText(keyString).color(JsonColor.SECONDARY));
                 helpJson.addExtra(BedrockJson.Space());
 
                 cmdDesc.addExtra(BedrockJson.Space());
                 cmdDesc.addExtra(BedrockJson.JsonWithText(keyString).color(JsonColor.SECONDARY));
-                if (!keyValueArgument.getKeyOnly()) {
+                if (option.hasParameter()) {
                     cmdDesc.addExtra(BedrockJson.Space());
                 }
             }
@@ -375,11 +375,10 @@ public abstract class Messages {
              * Argument placeholder
              */
 
-            if (!(argument instanceof KeyValueArgument)
-                    || !((KeyValueArgument) argument).getKeyOnly()) {
-                BedrockJson runtimePlaceholder = null;
+            if (!(argument instanceof Option) || ((Option) argument).hasParameter()) {
+                BedrockJson runtimePlaceholder;
 
-                if (argument instanceof KeyValueArgument) {
+                if (argument instanceof Option) {
                     runtimePlaceholder = BedrockJson.JsonWithText((argument.isOptional()
                             ? argument.getRuntimePlaceholder() + "]"
                             : argument.getRuntimePlaceholder()))
